feat: adicionando uma estrutura de model e controllers
