+++ Library/back/.env.example
@@ -1,0 +1,7 @@
+// .env
+DB_USER=isabellaoliveira
+DB_PASS=isa2004tb
+SECRET=EUVOTOMARUMTACACA
+SENDGRID_API_KEY=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
+
+
